+++ examples/models/models_first_person_maze.cpp
@@ -75,7 +75,7 @@ int main(void)
             for (int x = 0; x < cubicmap.width; x++)
             {
                 if ((mapPixels[y*cubicmap.width + x].r == 255) &&       // Collision: white pixel, only check R channel
-                    (playerPos.CheckCollisionCircleRec(playerRadius,
+                    (playerPos.CheckCollisionCircle(playerRadius,
                     (Rectangle){ mapPosition.x - 0.5f + x*1.0f, mapPosition.z - 0.5f + y*1.0f, 1.0f, 1.0f })))
                 {
                     // Collision detected, reset camera position